Add xlink namespace for SVG <image> elements
Safari won't show `<image>` elements in an SVG with the `href` attribute. This needs to be `xlink:href`, so this commit adds the `xlink` namepsace when adding the image attributes, so that these embedded images are rendered properly. It also tidies up a bit, making `XMLNS` and `XLINK` constants, so that they can be re-used throughout the component.

Fixes the issues outlined in this Trello card: https://trello.com/c/H63CKopL